\name{businessPubMed-package}
\alias{businessPubMed-package}
\alias{businessPubMed}
\docType{package}
\title{
\packageTitle{businessPubMed}
}
\description{
\packageDescription{businessPubMed}
}
\author{
\packageAuthor{businessPubMed}
Maintainer: \packageMaintainer{businessPubMed}
}
\references{
\url{
http://www.biotechworld.it/bioinf/2016/01/21/scraping-pubmed-data-via-easypubmed-xml-and-regex-in-r-for-a-targeting-campaign/
}
}

